Add scoreboard
