+++ index.html
@@ -9,7 +9,9 @@
   <div id="currentNumber"></div>
   <div><p id="currentConversion">Binary: </p></div>
   <div><p id="currentScore"></p></div>
+
   <canvas id="canvas" width="1000" height="600"></canvas>
+  <div><p id="scoreBoard"></p></div>
   <script src="main.bundle.js"></script>
 </body>
 </html>
