+++ lib/game.js
@@ -3,6 +3,7 @@ var Snake = require("./snake")
 var UserInput = require("./userInput")
 var Food = require("./food")
 var Num = require("./num")
+var scoreBoard = require("./scoreBoard")
 var $ = require("jquery");
 
 class Game {
@@ -16,6 +17,7 @@ class Game {
     this.speed = 2;
     this.currentNumber = new Num(10);
     this.score = 0;
+    this.scoreBoard = new scoreBoard();
   }
 
   setup() {
@@ -23,6 +25,7 @@ class Game {
     this.replenishFood();
     this.displayNumber();
     this.snake.startLength();
+    this.render.displayScoreBoard(this.scoreBoard.getHighScores());
     console.log(this.currentNumber.decimal, this.currentNumber.binary);
     console.log("to eat: ", this.currentNumber.nextBit())
   }
