+++ lib/index.js
@@ -16,26 +16,7 @@ requestAnimationFrame(function test() {
   }
     else {
       game.render.displayGameOver();
-      updateHighScores(getHighScores(), game.score);
+      game.scoreBoard.updateHighScores(game.score);
+      game.render.displayScoreBoard(game.scoreBoard.getHighScores());
   }
 });
-
-function updateHighScores(scores, currentScore) {
-    scores.push(currentScore);
-    var sortedScores = scores.sort(function(a, b) {
-      return parseInt(b) - parseInt(a);
-    });
-    localStorage.setItem("highScores", sortedScores.slice(0, 9))
-    console.log(localStorage.highScores)
-  }
-
-
-function getHighScores() {
-  if (!localStorage.highScores) {
-    localStorage.setItem("highScores", "0");
-    return [localStorage.getItem("highScores")];
-  }
-  else {
-    return localStorage.getItem("highScores").split(",");
-  }
-}
