+++ lib/render.js
@@ -59,6 +59,10 @@ class Render {
   displayScore(score) {
     $("#currentScore").text("Score: " + score);
   }
+
+  displayScoreBoard(highScores) {
+    $("#scoreBoard").text(highScores);
+  }
 }
 
 module.exports = Render;
